# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: Full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
...
  internal_features_users_transport_http.CrateUserReqDTO:
    properties:
      full_name:
        example: Name Name Name
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+380999999"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http.CreateUserResponse:
    properties:
      full_name:
        example: Name Name
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+380999999"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.GetUserResponse:
    properties:
      full_name:
        example: Name Name
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+380999999"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserRequest:
    properties:
      full_name:
        example: Name Name Name
        type: string
      phone_number:
        example: "+3809999999"
        type: string
    type: object
  internal_features_users_transport_http.PatchUserResponse:
    properties:
      full_name:
        example: Name Name
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+380999999"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.UsersDTOREsponse:
    properties:
      full_name:
        example: Name Name
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+380999999"
        type: string
      version:
        example: 3
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: TDO REST-API schema
  title: todoapp (TDO)
  version: "1.0"
paths:
...
  /users:
    get:
      description: Отримати список користувачів
      parameters:
      - description: розмір сторінки з користувачами
        in: query
        name: limit
        type: integer
      - description: зсув сторінки з користувачами
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успішне отрмання користувачів
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http.UsersDTOREsponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Отримати список користувачів
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Створити нового користувача в системі
      parameters:
      - description: CreateUser
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.CrateUserReqDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Успішно створений користувач
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Створити користувача
      tags:
      - users
  /users/{id}:
    delete:
      description: Видалення користувача
      parameters:
      - description: ID юзера якого слід видалити
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успішно видаленя користувача
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Видалення користувача
      tags:
      - users
    get:
      description: Отримати параметри користувача
      parameters:
      - description: ID юзера якого слід знайти
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успішно віднайдений користувач
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Отримати параметри користувача
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Змінти параметри користувача
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      - description: PatchUser тіло запиту
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успішно змінений користувач
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_DimaKirejko_todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Змінти параметри користувача
      tags:
      - users
swagger: "2.0"